Use matrix displayname if available
Use the matrix displayname instead of the matrix ID. The matrix ID is
often less readable, especially when the sender belongs to another
bridge which names puppets using raw UUIDs (such as
matrix-appserver-minecraft) which are a mouthful for Mumble's TTS!

The displayname is retrieved using getProfileInfo() from the matrix
bridge onEvent callback, which is now made async to handle the await,
and passed through to the Murmur sendMessage as a new argument. We
reference the bridge variable in the outer function as the callback
shouldn't get used until after it is initialised. In the unlikely event
that the getProfileInfo call fails, the exception is caught and we fall
back to using the sender's Matrix ID as before.